move vcpu & vm abstractions to new crate - vm-vcpu
As part of these change, a few functions needed to become public.
Updated the vm-device dependency to a fixed one to ensure that we do not
break the build when changing interfaces in vm-device.
