+++ src/vm-vcpu/src/lib.rs
@@ -1,0 +1,4 @@
+#![cfg(target_arch = "x86_64")]
+
+pub mod vcpu;
+pub mod vm;
