+++ src/vm-vcpu/src/vcpu/mpspec.rs
