+++ src/vm-vcpu/src/vcpu/mptable.rs
@@ -10,12 +10,9 @@ use std::mem;
 use std::result;
 use std::slice;
 
+use crate::vcpu::mpspec;
 use libc::c_char;
-
 use vm_memory::{Address, ByteValued, Bytes, GuestAddress, GuestMemory, GuestMemoryMmap};
-
-use crate::mpspec;
-
 // This is a workaround to the Rust enforcement specifying that any implementation of a foreign
 // trait (in this case `ByteValued`) where:
 // *    the type that is implementing the trait is foreign or
